move contexts into their own folder
